refactoring and comments added
