--- back/src/auth/index.js
@@ -1,10 +1,0 @@
-const { signUpRouter, signInRouter } = require('./auth');
-
-function registerAuthRoutes(app) {
-    app.use('/auth', signUpRouter);
-    app.use('/auth', signInRouter);
-};
-
-module.exports = {
-    registerAuthRoutes
-}
