+++ back/src/routes/auth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const signUpRouter = express.Router();
 const signInRouter = express.Router();
-const auth = require('./handlers/auth');
+const auth = require('../handlers/auth');
 
 signUpRouter.post('/signup', auth.signUp);
 signInRouter.post('/signin', auth.signIn);
