+++ back/src/routes/index.js
@@ -1,10 +1,11 @@
-const { registerAuthRoutes } = require('../auth');
 const { gifRouter } = require('./gif');
+const { signUpRouter, signInRouter } = require('./auth');
 
 function registerRoutes(app) {
 
     // authentification routes
-    registerAuthRoutes(app);
+    app.use('/auth', signUpRouter);
+    app.use('/auth', signInRouter);
 
     // entity routes
     app.use("/gifs", gifRouter);
