+++ front/src/index.js
@@ -3,10 +3,14 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import './styles/index.css';
 
-// Redux: Store creation, entirely available to the application.
+// Redux: Calling createStore and Provider methods to be able to create and provide the store.
 import { createStore } from 'redux';
 import { Provider } from  'react-redux';
+
+// Redux: Calling reducers.
 import allReducers from './reducers';
+
+// Redux: Store creation, entirely available to the application.
 const  store  =  createStore(allReducers)
 
 ReactDOM.render(
