+++ front/src/reducers/authReducer.js
@@ -1,6 +1,7 @@
+// Redux: Reducer from the authentication containers (SignUp/SignIn)
 export default function (state={}, action) {
     switch (action.type) {
-        case "CREATE_SESSION":
+        case "AUTH":
             return {...state, token: action.token, msg: action.message};
         default:
             return state;
