+++ front/src/reducers/index.js
@@ -2,6 +2,7 @@ import { combineReducers } from 'redux';
 import authReducer from './authReducer';
 import regReducer from './regReducer';
 
+// Redux: Combination of reducers to create the global reducer.
 const allReducers = combineReducers({
     auth: authReducer,
     reg: regReducer
