+++ front/src/reducers/regReducer.js
@@ -1,3 +1,4 @@
+// Redux: Reducer from the registration container (SignUp)
 export default function (state={}, action) {
     switch (action.type) {
         case "REGISTER_USER":
